pyproject.toml
[project]
name = "mox-cu-web3-favorites"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "vyper>=0.4.3",
    "web3>=7.14.0",
]
